Introduce a `page_index`.
Page index behaves similarly to the main index, except that it is to be
used inside pages (as opposed to articles) with the page hierarchy
plugins.

That is, given the following structure:

```
content/pages/
└─ project.rst
   project/
   ├─ page1.rst
   └─ page2.rst
```

if `project.rst` is assigned the `page_index` template, it will list
`page1.rst` and `page2.rst`.

Signed-off-by: JP-Ellis <coujellis@gmail.com>